4.22

+++ osa4/blogilista/controllers/blogs.js
@@ -24,7 +24,8 @@ blogsRouter.post('/', async (request, response, next) => {
   if (!token || !decodedToken.id) {
     return response.status(401).json({ error: 'token missing or invalid' })
   }
-  const user = await User.findById(decodedToken.id)
+  //const user = await User.findById(decodedToken.id)
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
@@ -43,8 +44,17 @@ blogsRouter.post('/', async (request, response, next) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  //const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  //const user = await User.findById(decodedToken.id)
+  const user = request.user
+
+  if ( blog.user.toString() === user.id.toString() ) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'invalid user or token missing or invalid' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response, next) => {
